feat: privacy implemented

+++ controllers/contactsControllers.js
@@ -2,15 +2,17 @@ import HttpError from "../helpers/HttpError.js";
 import ctrlWrapper from "../helpers/ctrlWrapper.js";
 import * as contactsService from "../services/contactsServices.js";
 
-export const getAllContacts = ctrlWrapper(async (_, res) => {
-  const result = await contactsService.listContacts();
+export const getAllContacts = ctrlWrapper(async (req, res) => {
+  const {_id: owner} = req.user;
+  const result = await contactsService.listContacts({owner});
 
   res.json(result);
 });
 
 export const getOneContact = ctrlWrapper(async (req, res) => {
   const {id} = req.params;
-  const result = await contactsService.getContactById(id);
+  const {_id: owner} = req.user;
+  const result = await contactsService.getContactById({_id: id, owner});
 
   if (!result) {
     throw HttpError(404, `Contact with id=${id} not found`);
@@ -21,8 +23,9 @@ export const getOneContact = ctrlWrapper(async (req, res) => {
 
 export const deleteContact = ctrlWrapper(async (req, res) => {
   const {id} = req.params;
+  const {_id: owner} = req.user;
 
-  const result = await contactsService.removeContact(id);
+  const result = await contactsService.removeContact({_id: id, owner});
 
   if (!result) {
     throw HttpError(404, `Contact with id=${id} not found`);
@@ -32,16 +35,18 @@ export const deleteContact = ctrlWrapper(async (req, res) => {
 });
 
 export const createContact = ctrlWrapper(async (req, res, next) => {
+  const {_id: owner} = req.user;
   const {name, email, phone} = req.body;
-  const result = await contactsService.addContact(name, email, phone);
+  const result = await contactsService.addContact(name, email, phone, owner);
 
   res.status(201).json(result);
 });
 
 export const updateContact = ctrlWrapper(async (req, res, next) => {
   const {id} = req.params;
+  const {_id: owner} = req.user;
 
-  const result = await contactsService.updContact(id, req.body);
+  const result = await contactsService.updContact({_id: id, owner}, req.body);
 
   if (!result) {
     throw HttpError(404, `Contact with id=${id} not found`);
@@ -52,9 +57,10 @@ export const updateContact = ctrlWrapper(async (req, res, next) => {
 
 export const updateStatusContact = ctrlWrapper(async (req, res) => {
   const {id} = req.params;
+  const {_id: owner} = req.user;
   const {favorite = false} = req.body;
 
-  const result = await contactsService.updContact(id, {favorite});
+  const result = await contactsService.updContact({_id: id, owner}, {favorite});
 
   if (!result) {
     throw HttpError(404, `Contact with id=${id} not found`);
